@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # Require this file using `require "spec_helper"` to ensure that it is only
@@ -11,18 +13,19 @@ require 'active_support/core_ext'
 require 'stator'
 
 RSpec.configure do |config|
-  config.treat_symbols_as_metadata_keys_with_true_values = true
   config.expect_with(:rspec) { |c| c.syntax = :should }
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
+
+  config.raise_errors_for_deprecations!
 
   NullDB.configure do |c|
     c.project_root = File.dirname(__FILE__)
   end
 
   ActiveRecord::Base.establish_connection(
-    :adapter => :nulldb,
-    :schema  => 'support/schema.rb'
+    adapter: :nulldb,
+    schema: 'support/schema.rb'
   )
 
   require 'support/models'
