@@ -2,13 +2,11 @@
 
 module Stator
   class Integration
-
     delegate :states,       to: :@machine
     delegate :transitions,  to: :@machine
     delegate :namespace,    to: :@machine
 
-    attr_reader :skip_validations
-    attr_reader :skip_transition_tracking
+    attr_reader :skip_validations, :skip_transition_tracking
 
     def initialize(machine, record)
       @machine = machine
@@ -36,6 +34,7 @@ module Stator
       return false unless @record.respond_to?(field_name)
       return false if @record.send(field_name).nil?
       return true if time.nil?
+
       @record.send(field_name) <= time
     end
 
@@ -63,7 +62,7 @@ module Stator
 
     # TODO: i18n
     def invalid_state!
-      @record.errors.add(@machine.field, "is not a valid state")
+      @record.errors.add(@machine.field, 'is not a valid state')
     end
 
     def invalid_transition!(was, is)
@@ -115,9 +114,7 @@ module Stator
       later_states = later_groups[later_group_key]
 
       # if the lowest timestamp is the same as the state's timestamp, evaluate based on state index
-      if later_states[0][:at] == state_at
-        return all_states.index(state) < all_states.index(later_states[0][:state])
-      end
+      return all_states.index(state) < all_states.index(later_states[0][:state]) if later_states[0][:at] == state_at
 
       false
     end
@@ -162,6 +159,5 @@ module Stator
 
       @record.send("#{field_name}=", (Time.zone || Time).now)
     end
-
   end
 end
