@@ -1,12 +1,10 @@
 # frozen_string_literal: true
 
 module Stator
-
   MAJOR       = 0
   MINOR       = 4
   PATCH       = 0
   PRERELEASE  = nil
 
-  VERSION = [MAJOR, MINOR, PATCH, PRERELEASE].compact.join(".")
-
+  VERSION = [MAJOR, MINOR, PATCH, PRERELEASE].compact.join('.')
 end
